internal: move derived tests to the unified macro infra
